chore: useVisualizationContext
