+++ packages/frontend/src/components/LightdashVisualization/VisualizationProvider/useVisualizationContext.tsx
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { Context, VisualizationContext } from '.';
+
+export function useVisualizationContext(): VisualizationContext {
+    const context = useContext(Context);
+    if (context === undefined) {
+        throw new Error(
+            'useVisualizationContext must be used within a VisualizationProvider',
+        );
+    }
+    return context;
+}
